eco2mix {leaflet.minicharts}	R Documentation

Electric production, consumption and exchanges of France

Description

eco2mix contains the electric production, consumption and exchanges of France from january 2010 to february 2017 and of 12 french regions from january 2013 to february 2017.

In addition to the total production, the table contains one column for each type of production. The table also contains the lattitude and longitude of the center of the regions.

eco2mixBalance is an extract of eco2mix that contains only exchanges between France and neighbouring countries, in a convenient format to represent flows on a map.

Usage

eco2mix

eco2mixBalance

Format

An object of class data.frame with 686 rows and 22 columns.

Author(s)

Francois Guillem francois.guillem@rte-france.com

References

http://www.rte-france.com/fr/eco2mix/eco2mix

[Package leaflet.minicharts version 0.5.4 Index]