popupArgs {leaflet.minicharts}	R Documentation

Options for popup generation

Description

This function simply returns a list of options to control the generation of popups.

Usage

popupArgs(showTitle = TRUE, showValues = TRUE, labels = NULL,
  supValues = NULL, supLabels = colnames(supValues), html = NULL,
  noPopup = FALSE, digits = NULL)

Arguments

showTitle
If TRUE layer id is displayed as title of popups.

showValues
If TRUE, values are displayed in popups

labels
Names of values. If NULL, column names of the data bound to a chart are used.

supValues
A data.frame containing additional values to display in popups.

supLabels
Names of the additional values.

html
Character vector containing custom html code for popups. You can use this parameter when you are not happy with the default popups.

noPopup
If TRUE, popups are not created.

digits
Max number of decimal digits to display for numeric values. If NULL, all digits are displayed.

Value

List containing options for popup generation

[Package leaflet.minicharts version 0.5.4 Index]